Update for sam stuff
Add in a script to make the table used for examples, this is sourced in latex.
Also, more importantly, needed to add some fixes to sam-gen.r. I guess I had added
in log-likelihood and forgot to agjust code sufficiently. I also
added in comments to the Rcpp code for sam and compressed a few declaration
lines.